test(P2): move pipeline-wrapper tests to their own file
Relocate the four _run_* wrapper tests out of tests/test_pipeline.py into
tests/test_pipeline_wrappers.py. Functionally identical; this just avoids an
append-collision with other branches that also add tests to test_pipeline.py,
so every audit PR merges into main cleanly and in any order.
